refactor: organize accelerator backends
