+++ AGENTS/experience_reports/1749784532_v1_Accelerator_Backend_Subfolder.md
@@ -1,0 +1,9 @@
+# Accelerator Backend Subfolder
+
+## Summary
+- Moved `c_backend.py` and `opengl_backend.py` into new `tensors/accelerator_backends/`.
+- Updated imports across package and tests.
+- Added `AcceleratorCoordinator` stub to unify buffer management.
+
+## Prompt History
+User: "pursue the conceptual flag for c and opengl backend with a subfolder in tensors..."
