+++ AGENTS/experience_reports/archive/1749769234_v1_Tools_Group_Removal.md
