+++ AGENTS/experience_reports/archive/1749772107_v1_Poetry_Migration.md
