+++ tensors/accelerator_backends/coordinator.py
@@ -1,0 +1,57 @@
+#!/usr/bin/env python3
+"""Backend-agnostic buffer coordinator for accelerator operations."""
+from __future__ import annotations
+
+try:
+    from AGENTS.tools.header_utils import ENV_SETUP_BOX
+    from typing import Any, Deque, Tuple
+    from collections import deque
+
+    from .c_backend import CTensorOperations
+    from .opengl_backend import OpenGLTensorOperations
+except Exception:
+    import sys
+    print(ENV_SETUP_BOX)
+    sys.exit(1)
+# --- END HEADER ---
+
+
+# ########## STUB: Accelerator Buffer Coordinator ##########
+# PURPOSE: Provide a unified interface that manages instruction chains and
+#          contiguous buffers for both the C and OpenGL backends.
+# EXPECTED BEHAVIOR: When implemented, this class will enqueue operations and
+#          dispatch them to either a thread-saturated C backend or a compute
+#          shader batch on the OpenGL backend. It abstracts buffer management so
+#          client code remains agnostic to the underlying accelerator.
+# INPUTS: Instruction objects describing tensor transformations and a selected
+#         backend implementation.
+# OUTPUTS: Finalized contiguous buffers ready for consumption by the caller.
+# KEY ASSUMPTIONS/DEPENDENCIES: Requires either ``CTensorOperations`` or
+#         ``OpenGLTensorOperations`` to be available. Assumes instruction chains
+#         can be executed sequentially on a worker thread or GPU queue.
+# TODO:
+#   - Implement an instruction representation for tensor operations.
+#   - Add worker thread or GPU dispatch logic.
+#   - Provide synchronization primitives to retrieve completed buffers.
+# NOTES: This is an architectural scaffold to pursue the conceptual flag
+#         unifying C and OpenGL execution paths.
+# ###########################################################################
+class AcceleratorCoordinator:
+    """Orchestrate C or OpenGL tensor backends via instruction queues."""
+
+    def __init__(self, backend: str = "c") -> None:
+        if backend == "c":
+            self.ops = CTensorOperations()
+        elif backend == "opengl":
+            self.ops = OpenGLTensorOperations()
+        else:
+            raise ValueError(f"Unknown backend: {backend}")
+        self.queue: Deque[Tuple[str, Tuple[Any, ...], dict[str, Any]]] = deque()
+
+    def enqueue(self, name: str, *args: Any, **kwargs: Any) -> None:
+        """Queue an operation by name with arguments."""
+        self.queue.append((name, args, kwargs))
+
+    def execute(self) -> Any:
+        """Execute all queued operations sequentially (stub)."""
+        raise NotImplementedError("AcceleratorCoordinator.execute not implemented")
